+++ src/parser/enrichers/class/monk/DeflectMissiles.ts
@@ -35,7 +35,7 @@ export default class DeflectMissiles extends DDBEnricherData {
         action: { name: "Deflect Missiles Attack", type: "class", rename: ["Deflect Missiles Attack"] },
         overrides: {
           addItemConsume: true,
-          itemConsumeTargetName: "Ki",
+          itemConsumeTargetName: this.is2014 ? "Ki" : "Monk's Focus",
         },
       },
     ];