+++ src/parser/enrichers/class/monk/EmptyBody.ts
@@ -37,7 +37,7 @@ export default class EmptyBody extends DDBEnricherData {
         overrides: {
           addSpellUuid: "Astral Projection",
           addItemConsume: true,
-          itemConsumeTargetName: "Ki",
+          itemConsumeTargetName: this.is2014 ? "Ki" : "Monk's Focus",
           itemConsumeValue: 8,
           data: {
             spell: {