+++ src/parser/enrichers/class/monk/SearingSunburst.ts
@@ -7,7 +7,7 @@ export default class SearingSunburst extends Generic {
     return {
       activationType: "action",
       itemConsumeValue: 0,
-      itemConsumeTargetName: "Ki",
+      itemConsumeTargetName: this.is2014 ? "Ki" : "Monk's Focus",
       addItemConsume: true,
       addScalingMode: "amount",
       addScalingFormula: "1",